// src/generated/graphql.tsx
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string };
  String: { input: string; output: string };
  Boolean: { input: boolean; output: boolean };
  Int: { input: number; output: number };
  Float: { input: number; output: number };
  DateTime: { input: any; output: any };
  Decimal: { input: any; output: any };
  GraphqlCursor: { input: any; output: any };
  GraphqlPoint: { input: any; output: any };
  JSON: { input: any; output: any };
  JSONObject: { input: any; output: any };
};
// ...
export type CustomerPage = {
  __typename?: "CustomerPage";
  cursor: Scalars["GraphqlCursor"]["output"];
  data: Array<Customer>;
};
// ...
export type Discount = {
  __typename?: "Discount";
  createdAt: Scalars["DateTime"]["output"];
  endDiscount: Scalars["DateTime"]["output"];
  id: Scalars["Int"]["output"];
  startDiscount: Scalars["DateTime"]["output"];
  type: DiscountType;
};

export enum DiscountType {
  Fixed = "fixed",
  FreeShipping = "freeShipping",
  Percentage = "percentage",
}

export type DomesticGovernorate = {
  __typename?: "DomesticGovernorate";
  arabicName: Scalars["String"]["output"];
  id: Scalars["Int"]["output"];
  name: Scalars["String"]["output"];
};

export type DomesticShipping = {
  __typename?: "DomesticShipping";
  base: Scalars["Decimal"]["output"];
  createdAt: Scalars["DateTime"]["output"];
  extra: Scalars["Decimal"]["output"];
  id: Scalars["Int"]["output"];
  lastModified: Scalars["DateTime"]["output"];
  post: Scalars["Decimal"]["output"];
  total: Scalars["Decimal"]["output"];
  vat: Scalars["Decimal"]["output"];
};

export type DomesticZone = {
  __typename?: "DomesticZone";
  governorateId: Scalars["Int"]["output"];
  id: Scalars["Int"]["output"];
  name: Scalars["String"]["output"];
  nameAr: Scalars["String"]["output"];
};

export type Employee = {
  __typename?: "Employee";
  commission: Scalars["Decimal"]["output"];
  createdAt: Scalars["DateTime"]["output"];
  currency: Scalars["String"]["output"];
  id: Scalars["String"]["output"];
  lastModified: Scalars["DateTime"]["output"];
  salary: Scalars["Decimal"]["output"];
  type: EmployeeType;
  user: User;
};
// ...
export type FinancialTransactionTotal = {
  __typename?: "FinancialTransactionTotal";
  count: Scalars["Int"]["output"];
  currency?: Maybe<Scalars["String"]["output"]>;
  total?: Maybe<Scalars["Decimal"]["output"]>;
};
// ...
export type Hub = {
  __typename?: "Hub";
  createdAt: Scalars["DateTime"]["output"];
  governorateId: Scalars["Int"]["output"];
  id: Scalars["Int"]["output"];
  lastModified: Scalars["DateTime"]["output"];
  location?: Maybe<Scalars["GraphqlPoint"]["output"]>;
  name: Scalars["String"]["output"];
  users: Array<User>;
};
// ...
export type InventoryRent = {
  __typename?: "InventoryRent";
  createdAt: Scalars["DateTime"]["output"];
  currency: Scalars["String"]["output"];
  deletedAt?: Maybe<Scalars["DateTime"]["output"]>;
  id: Scalars["Int"]["output"];
  lastBill?: Maybe<Scalars["DateTime"]["output"]>;
  lastModified: Scalars["DateTime"]["output"];
  merchantId: Scalars["Int"]["output"];
  pricePerUnit: Scalars["Decimal"]["output"];
  sqaureMeter?: Maybe<Scalars["Decimal"]["output"]>;
  startDate: Scalars["DateTime"]["output"];
  type: InventoryRentType;
};
// ...
export type ItemPage = {
  __typename?: "ItemPage";
  cursor: Scalars["GraphqlCursor"]["output"];
  data: Array<Item>;
};
// ...
export type ItemVariant = {
  __typename?: "ItemVariant";
  createdAt: Scalars["DateTime"]["output"];
  deletedAt?: Maybe<Scalars["DateTime"]["output"]>;
  id: Scalars["Int"]["output"];
  imageUrl: Scalars["String"]["output"];
  isEnabled: Scalars["Boolean"]["output"];
  item: Item;
  itemId: Scalars["Int"]["output"];
  lastModified: Scalars["DateTime"]["output"];
  merchant: Merchant;
  merchantId: Scalars["Int"]["output"];
  merchantSku: Scalars["String"]["output"];
  name?: Maybe<Scalars["String"]["output"]>;
  price: Scalars["Decimal"]["output"];
  selectedOptions: Array<ItemVariantOption>;
  shopifyId?: Maybe<Scalars["String"]["output"]>;
  sku: Scalars["String"]["output"];
  weight?: Maybe<Scalars["Decimal"]["output"]>;
};
// ...
export type MerchantDomesticShipping = {
  __typename?: "MerchantDomesticShipping";
  createdAt: Scalars["DateTime"]["output"];
  domesticShipping?: Maybe<DomesticShipping>;
  domesticShippingId: Scalars["Int"]["output"];
  governorate: DomesticGovernorate;
  governorateId: Scalars["Int"]["output"];
  id: Scalars["Int"]["output"];
  lastModified: Scalars["DateTime"]["output"];
  merchantId: Scalars["Int"]["output"];
  merchantJoined?: Maybe<Merchant>;
  orderType: OrderType;
};
// ...
export type Query = {
  __typename?: "Query";
  CourierSheet: CourierSheet;
  ballot: Ballot;
  box: Box;
  calculateFinancialTransactionsTotal: FinancialTransactionTotal;
  countInventoryRentTransaction: Scalars["Int"]["output"];
  createMerchantToken: Scalars["JSONObject"]["output"];
  discount: Discount;
  employee: Employee;
  findAllDomesticGovernorates: Array<DomesticGovernorate>;
  findAllHubs: Array<Hub>;
  findAllZones: Array<DomesticZone>;
  findCustomer: CustomerPage;
  findItemsWithShopifyId: Array<Item>;
  findMerchantDomesticShippings: Array<MerchantDomesticShipping>;
  findMerchantInventoryRent?: Maybe<InventoryRent>;
  findOneHub: Hub;
  findOneInventory: Inventory;
  findOneMerchant: Merchant;
  findOneOrder?: Maybe<Order>;
  findOneReturnPackage?: Maybe<ReturnPackage>;
  findOneZone: DomesticZone;
  findRoles: Array<Role>;
  findShop?: Maybe<Merchant>;
  findShopifyOrders: Array<Order>;
  findSimilarAddresses: Array<AddressSimilarityItem>;
  findSingleMerchantDomesticShipping: MerchantDomesticShipping;
  inventoryHistory: InventoryItemHistory;
  inventoryRentTransaction: InventoryRentTransaction;
  inventorySupport: InventorySupport;
  itemInBox: ItemInBox;
  itemReturnHistory: ItemReturnHistory;
  listener: Listener;
  listeners: Array<Listener>;
  merchantVisit: MerchantVisit;
  merchantVisitHistory: MerchantVisitHistory;
  myCourierSheet: CourierSheet;
  orderItem: OrderItem;
  orderRelation: OrderRelation;
  paginateAddresses: MerchantCustomerAddressPage;
  paginateBallots: BallotPage;
  paginateBoxes: BoxPage;
  paginateCourierCollectionTransactions: FinancialTransactionPage;
  paginateCourierSheets: CourierSheetPage;
  paginateCouriers: UserPage;
  paginateExpenses: ExpensesPage;
  paginateFinancialAccounts: FinancialAccountsPage;
  paginateFinancialTransaction: FinancialTransactionPage;
  paginateHubs: HubsPage;
  paginateInventories: InventoryPage;
  paginateInventoryRentTransaction: InventoryRentTransactionPage;
  paginateInventoryReturns: InventoryReturnPage;
  paginateItemHistory: ItemHistoryPage;
  paginateItemInBox: ItemInBoxPage;
  paginateItemReturns: ItemReturnPage;
  paginateItemVariants: ItemVariantPage;
  paginateItems: ItemPage;
  paginateMerchantPaymentTransactions: FinancialTransactionPage;
  paginateMerchants: MerchantPage;
  paginateOrderHistory: OrderHistoryPage;
  paginateOrderTransactionsHistory: FinancialTransactionPage;
  paginateOrders: OrderPage;
  paginateOrdersInInventory: OrderPage;
  paginateRacks: RacksPage;
  paginateReturnPackageHistory: ReturnPackageHistoryPage;
  paginateReturnPackages: ReturnPackagePage;
  paginateUsers: UserPage;
  request: Request;
  requestStatusHistory: RequestStatusHistory;
  returnPackageHistory: ReturnPackageHistory;
  role: Role;
  sendToEvent: EventListenObject;
  sheetOrder: SheetOrder;
  sheetOrderStatusHistory: SheetOrderStatusHistory;
  shopifySession: ShopifySession;
  sumInventoryRentTransaction: Scalars["Decimal"]["output"];
  testSheet: Scalars["Boolean"]["output"];
  userState: Scalars["Int"]["output"];
  webFindOneOrder?: Maybe<Order>;
  webPaginateOrders: OrderPage;
};
// ...
export type ReturnPackage = {
  __typename?: "ReturnPackage";
  count: Scalars["Int"]["output"];
  courier?: Maybe<User>;
  courierId?: Maybe<Scalars["String"]["output"]>;
  createdAt: Scalars["DateTime"]["output"];
  hubId: Scalars["Int"]["output"];
  id: Scalars["Int"]["output"];
  inventory?: Maybe<Inventory>;
  lastModified: Scalars["DateTime"]["output"];
  merchant?: Maybe<Merchant>;
  sku: Scalars["String"]["output"];
  status: ReturnPackageStatus;
  toInventoryId?: Maybe<Scalars["Int"]["output"]>;
  toMerchantId?: Maybe<Scalars["Int"]["output"]>;
  type: ReturnPackageType;
};
// ...
export type Role = {
  __typename?: "Role";
  createdAt: Scalars["DateTime"]["output"];
  description: Scalars["String"]["output"];
  id: Scalars["Int"]["output"];
  lastModified: Scalars["DateTime"]["output"];
  name: Scalars["String"]["output"];
  type: RoleType;
};
// ...
export type Ballot = {
  __typename?: "ballot";
  boxes: Array<Box>;
  createdAt: Scalars["DateTime"]["output"];
  deletedAt?: Maybe<Scalars["DateTime"]["output"]>;
  id: Scalars["Int"]["output"];
  lastModified: Scalars["DateTime"]["output"];
  level: Scalars["Int"]["output"];
  merchant?: Maybe<Merchant>;
  merchantId?: Maybe<Scalars["Int"]["output"]>;
  name: Scalars["String"]["output"];
  rack: Rack;
  rackId: Scalars["Int"]["output"];
};

export type Box = {
  __typename?: "box";
  ballot: Ballot;
  ballotId: Scalars["Int"]["output"];
  createdAt: Scalars["DateTime"]["output"];
  deletedAt?: Maybe<Scalars["DateTime"]["output"]>;
  id: Scalars["Int"]["output"];
  lastModified: Scalars["DateTime"]["output"];
  merchant?: Maybe<Merchant>;
  merchantId?: Maybe<Scalars["Int"]["output"]>;
  name: Scalars["String"]["output"];
};

export type CompleteMerchantPaymentInput = {
  allTransactions?: InputMaybe<Scalars["Boolean"]["input"]>;
  description?: InputMaybe<Scalars["String"]["input"]>;
  transactionIds: Array<Scalars["Int"]["input"]>;
};

export type CourierSheet = {
  __typename?: "courierSheet";
  createdAt: Scalars["DateTime"]["output"];
  id: Scalars["Int"]["output"];
  lastModified: Scalars["DateTime"]["output"];
  orderCount: Scalars["Int"]["output"];
  sheetOrders?: Maybe<Array<SheetOrder>>;
  status: CourierSheetStatus;
  user: User;
  userId: Scalars["String"]["output"];
  userInfo: User;
};
// ...
export type Inventory = {
  __typename?: "inventory";
  createdAt: Scalars["DateTime"]["output"];
  governorateId?: Maybe<Scalars["Int"]["output"]>;
  hubId?: Maybe<Scalars["Int"]["output"]>;
  id: Scalars["Int"]["output"];
  inventorySupports: Array<InventorySupport>;
  lastModified: Scalars["DateTime"]["output"];
  location: Scalars["GraphqlPoint"]["output"];
  module: Scalars["Int"]["output"];
  name: Scalars["String"]["output"];
  racks: Array<Rack>;
};
// ...
export type Item = {
  __typename?: "item";
  createdAt: Scalars["DateTime"]["output"];
  currency: Scalars["String"]["output"];
  description?: Maybe<Scalars["String"]["output"]>;
  id: Scalars["Int"]["output"];
  imageUrl?: Maybe<Scalars["String"]["output"]>;
  itemVariants: Array<ItemVariant>;
  lastModified: Scalars["DateTime"]["output"];
  merchantId: Scalars["Int"]["output"];
  name: Scalars["String"]["output"];
  shopifyId?: Maybe<Scalars["String"]["output"]>;
};
// ...
export type Merchant = {
  __typename?: "merchant";
  address?: Maybe<CustomerAddress>;
  addressId?: Maybe<Scalars["Int"]["output"]>;
  bankName?: Maybe<Scalars["String"]["output"]>;
  bankNumber?: Maybe<Scalars["String"]["output"]>;
  createdAt: Scalars["DateTime"]["output"];
  currency: Scalars["String"]["output"];
  financialAccount: FinancialAccount;
  id: Scalars["Int"]["output"];
  includesVat: Scalars["Boolean"]["output"];
  inventoryRent?: Maybe<InventoryRent>;
  lastModified: Scalars["DateTime"]["output"];
  name: Scalars["String"]["output"];
  overShipping: Scalars["Decimal"]["output"];
  owner?: Maybe<User>;
  ownerId?: Maybe<Scalars["String"]["output"]>;
  shopifyShop?: Maybe<Scalars["String"]["output"]>;
  taxId?: Maybe<Scalars["String"]["output"]>;
  threshold: Scalars["Decimal"]["output"];
  users: Array<User>;
  webToken?: Maybe<Scalars["String"]["output"]>;
  webTokenExpiration: Scalars["DateTime"]["output"];
};
// ...
export type ProductListQuery = {
  __typename?: "Query";
  paginateItems: {
    __typename?: "ItemPage";
    cursor: any;
    data: Array<{
      __typename?: "item";
      id: number;
      merchantId: number;
      name: string;
      shopifyId?: string | null;
      description?: string | null;
      currency: string;
      imageUrl?: string | null;
      createdAt: any;
      lastModified: any;
      itemVariants: Array<{
        __typename?: "ItemVariant";
        id: number;
        sku: string;
        name?: string | null;
        shopifyId?: string | null;
        merchantSku: string;
        merchantId: number;
        itemId: number;
        isEnabled: boolean;
        imageUrl: string;
        price: any;
        weight?: any | null;
        createdAt: any;
        lastModified: any;
        deletedAt?: any | null;
      }>;
    }>;
  };
};